+++ lab_1/task_1/key.json
@@ -1,0 +1,3 @@
+{
+    "key" : "блинница"
+}
