--- lab_1/task_1/key.txt
@@ -1,1 +1,0 @@
-блинница
