+++ lab_1/task_1/text_encryptor.py
@@ -2,48 +2,7 @@ import argparse
 from os import path
 
 
-def read_file(path_to_file) -> str:
-    """
-    Читает содержимое файла и возвращает его в верхнем регистре.
-
-    Args:
-        path_to_file (str): Путь к файлу, который нужно прочитать.
-
-    Returns:
-        str: Содержимое файла в верхнем регистре.
-
-    Raises:
-        FileNotFoundError: Если файл не найден.
-        Exception: Если произошла другая ошибка при чтении файла.
-    """
-    try: 
-        with open(path_to_file, "r", encoding="utf-8") as file:
-            return file.read().strip().upper()
-    except FileNotFoundError as e:
-        raise FileNotFoundError(f"File was not found: {e}")
-    except Exception as e:
-        raise Exception(f"error when reading a file: {e}")
-
-
-def write_file(path_to_file, text) -> None:
-    """
-    Записывает текст в указанный файл.
-
-    Args:
-        path_to_file (str): Путь к файлу, в который нужно записать текст.
-        text (str): Текст, который нужно записать в файл.
-
-    Raises:
-        Exception: Если произошла ошибка при записи в файл.
-    """
-    try: 
-        with open(path_to_file, "w", encoding="utf-8") as file:
-            file.write(text)
-    except Exception as e:
-        raise Exception(f"error when writing a file: {e}")
-    
-
-def arg_parser() -> argparse.Namespace:
+def arg_parser_task_1() -> argparse.Namespace:
     """
     Парсинг аргументов командной строки.
 
@@ -55,7 +14,7 @@ def arg_parser() -> argparse.Namespace:
     """
     parser = argparse.ArgumentParser(description='The program for encrypting messages using the Vigener method')
     parser.add_argument("path_to_text", type=str, help="the path to the encrypted text")
-    parser.add_argument("path_to_key", type=str, help="the path to the encryption key")
+    parser.add_argument("path_to_key", type=str, help="the path to the encryption key (.json)")
     parser.add_argument("path_to_ciphertext", type=str, help="the path to the ciphertext storage location")
     args = parser.parse_args()
     if not path.exists(args.path_to_text):
@@ -122,7 +81,7 @@ def decrypt(chipertext, key) -> str:
 
     Raises:
         Exception: Если произошла ошибка при дешифровке.
-    """
+    """ 
     k_index = 0
     k_lenght = len(key) 
     dechipertext = ''
@@ -134,23 +93,3 @@ def decrypt(chipertext, key) -> str:
     except Exception as e:
         raise Exception(f"decripting error: {e}")
     return dechipertext
-
-
-def main():
-    """
-    Основная точка входа программы.
-
-    Читает аргументы командной строки, шифрует текст и сохраняет результат.
-    """
-    try:
-        paths = arg_parser()
-        text = read_file(paths.path_to_text)
-        key = read_file(paths.path_to_key)
-        print(key)
-        chipertext = encript(text, key)
-        write_file(paths.path_to_ciphertext, chipertext)
-    except Exception as e:
-        raise Exception(f"program error: {e}")
-
-if __name__ == "__main__":
-    main()
